Rename StatisticsCalculator -> CoTermsProcessor.
