--- indexing/src/main/java/uk/ac/ebi/quickgo/index/annotation/coterms/StatisticsCalculator.java
@@ -1,70 +1,0 @@
-package uk.ac.ebi.quickgo.index.annotation.coterms;
-
-import com.google.common.base.Preconditions;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicLong;
-import org.springframework.batch.item.ItemProcessor;
-
-/**
- * For a GO Term retrieve all the co-occurring terms together with the statistics related to that co-occurrence.
- *
- * @author Tony Wardell
- * Date: 26/11/2015
- * Time: 11:59
- * Created with IntelliJ IDEA.
- *
- */
-public class StatisticsCalculator implements ItemProcessor<String, List<CoTerm>> {
-
-    private final CoTermsAggregationWriter aggregator;
-
-    /**
-     *
-     * @param aggregator holds the data for co-occurring terms.
-     */
-    public StatisticsCalculator(CoTermsAggregationWriter aggregator) {
-        Preconditions.checkArgument(aggregator != null, "The aggregator instance passed to the Statistics Calculator " +
-                "constructor cannot be null");
-        this.aggregator = aggregator;
-    }
-
-    /**
-     * For the passed in GO Term id, find the list of co-occurring terms and calculate CoTerm instances.
-     * @param goTerm the GO Term id for which co-occurring statistics will be calculated, co-occurring term list of
-     * objects will be returned containing data and statistics.
-     * @return a list of CoTerm objects. Each object represents a single permutation of two GO Terms that are
-     * used to annotate the same gene product, and the statistics about that permutation.
-     */
-    @Override
-    public List<CoTerm> process(String goTerm) {
-        return createCoTermsForSelectedTerm(goTerm).highestSimilarity();
-    }
-
-    /**
-     * Create a CoTermsForSelectedTerm instance for each compared term.
-     * @param goTerm The GO Term for which the co-occurrence statistics will be calculated.
-     * @return CoTermsForSelectedTerm instance with co-occurring statistics calculated for every
-     * co-occurring term.
-     */
-    private CoTermsForSelectedTerm createCoTermsForSelectedTerm(String goTerm) {
-
-        Preconditions
-                .checkArgument(null != goTerm, "Target GO term id passed to createCoTermsForSelectedTerm should not " +
-                        "be null");
-
-        Map<String, AtomicLong> coTermsForTarget = aggregator.getCoTermsAndCounts(goTerm);
-        CoTermsForSelectedTerm.Builder coTermsBuilder = new CoTermsForSelectedTerm.Builder()
-                .setTotalNumberOfGeneProducts(aggregator.getTotalOfAnnotatedGeneProducts())
-                .setSelected(aggregator.getGeneProductCountForGoTerm(goTerm));
-
-        for (String comparedTerm : coTermsForTarget.keySet()) {
-            coTermsBuilder.addCoTerm(new CoTerm.Builder().setTarget(goTerm).setComparedTerm
-                    (comparedTerm)
-                    .setCompared(aggregator.getGeneProductCountForGoTerm(comparedTerm))
-                    .setTogether(coTermsForTarget.get(comparedTerm).get()).createCoTerm());
-        }
-        return coTermsBuilder.build();
-
-    }
-}
